api/requirements.txt
azure-functions==1.20.0
pyodbc==5.1.0
azure-servicebus==7.12.2
azure-communication-sms==1.0.1
azure-communication-email==1.0.0
python-tds==1.12.0
certifi==2025.6.2
tzdata==2024.1
